Refactor out large classes
