+++ src/humanPose/draw.js
@@ -1,632 +1,12 @@
 import * as THREE from '../../thirdPartyCode/three/three.module.js';
-import {JOINTS, JOINT_CONNECTIONS, JOINT_PUBLIC_DATA_KEYS, getJointNodeInfo} from './utils.js';
-import {annotateHumanPoseRenderer} from './rebaScore.js';
+import {JOINTS, JOINT_PUBLIC_DATA_KEYS, getJointNodeInfo} from './utils.js';
 import {SpaghettiMeshPath} from './spaghetti.js';
+import {HumanPoseRenderer} from './HumanPoseRenderer.js';
+import {HumanPoseRenderInstance} from './HumanPoseRenderInstance.js';
+import {RENDER_CONFIDENCE_COLOR, MAX_POSE_INSTANCES} from './constants.js';
 
 let humanPoseAnalyzer;
 const poseRenderers = {};
-
-const SCALE = 1000; // we want to scale up the size of individual joints, but not apply the scale to their positions
-const RENDER_CONFIDENCE_COLOR = false;
-
-const COLOR_BASE = new THREE.Color(0, 0.5, 1);
-const COLOR_RED = new THREE.Color(1, 0, 0);
-const COLOR_YELLOW = new THREE.Color(1, 1, 0);
-const COLOR_GREEN = new THREE.Color(0, 1, 0);
-
-const MAX_POSE_INSTANCES = 1024 * 2;
-const JOINTS_PER_POSE = Object.keys(JOINTS).length;
-const BONES_PER_POSE = Object.keys(JOINT_CONNECTIONS).length;
-
-const JOINT_TO_INDEX = {};
-for (const [i, jointId] of Object.values(JOINTS).entries()) {
-    JOINT_TO_INDEX[jointId] = i;
-}
-
-const BONE_TO_INDEX = {};
-for (const [i, boneName] of Object.keys(JOINT_CONNECTIONS).entries()) {
-    BONE_TO_INDEX[boneName] = i;
-}
-
-/**
- * @param {THREE.InstancedBufferAttribute} instancedBufferAttribute
- * @param {number} slot - offset within iba in units of slots
- * @param {number} slotWidth - width of each slot in units of iba's item size
- * @return {TypedArray} array of length `slotWidth * itemSize`
- */
-function getSliceSlot(instancedBufferAttribute, slot, slotWidth) {
-    const itemSize = instancedBufferAttribute.itemSize;
-    const start = slot * slotWidth * itemSize;
-    return instancedBufferAttribute.array.slice(start, start + slotWidth * itemSize);
-}
-
-function setSliceSlot(instancedBufferAttribute, slot, slotWidth, items) {
-    const itemSize = instancedBufferAttribute.itemSize;
-    const start = slot * slotWidth * itemSize;
-
-    unionUpdateRange(instancedBufferAttribute, start, slotWidth * itemSize);
-
-    for (let i = 0; i < slotWidth * itemSize; i++) {
-        instancedBufferAttribute.array[start + i] = items[i];
-    }
-}
-
-function unionUpdateRange(instancedBufferAttribute, offset, count) {
-    if (instancedBufferAttribute.updateRange.count === -1) {
-        instancedBufferAttribute.updateRange.offset = offset;
-        instancedBufferAttribute.updateRange.count = count;
-        return;
-    }
-    let curMin = instancedBufferAttribute.updateRange.offset;
-    let curMax = curMin + instancedBufferAttribute.updateRange.count;
-    let plusMin = offset;
-    let plusMax = offset + count;
-    let newMin = Math.min(curMin, plusMin);
-    let newMax = Math.max(curMax, plusMax);
-    instancedBufferAttribute.updateRange.offset = newMin;
-    instancedBufferAttribute.updateRange.count = newMax - newMin;
-}
-
-/**
- * Manager of multiple HumanPoseRenderInstances within two instanced meshes
- */
-class HumanPoseRenderer {
-    /**
-     * @param {THREE.Material} material - Material for all instanced meshes
-     */
-    constructor(material, maxInstances) {
-        this.container = new THREE.Group();
-        // A stack of free instance slots (indices) that a PoseRenderInstance
-        // can reuse
-        this.freeInstanceSlots = [];
-        this.nextInstanceSlot = 0;
-        this.maxInstances = maxInstances;
-        this.createMeshes(material);
-    }
-
-    /**
-     * Creates all THREE.Meshes representing the spheres/joint balls of the
-     * pose
-     * @param {THREE.Material} material - Material for all instanced meshes
-     */
-    createMeshes(material) {
-        const geo = new THREE.SphereGeometry(0.03 * SCALE, 12, 12);
-
-        this.jointsMesh = new THREE.InstancedMesh(
-            geo,
-            material,
-            JOINTS_PER_POSE * this.maxInstances,
-        );
-        // Initialize instanceColor
-        this.jointsMesh.setColorAt(0, COLOR_BASE);
-        this.jointsMesh.count = 0;
-        this.jointsMesh.instanceMatrix.setUsage(THREE.DynamicDrawUsage);
-        this.jointsMesh.instanceColor.setUsage(THREE.DynamicDrawUsage);
-
-        this.container.add(this.jointsMesh);
-
-        const geoCyl = new THREE.CylinderGeometry(0.01 * SCALE, 0.01 * SCALE, SCALE, 3);
-        this.bonesMesh = new THREE.InstancedMesh(
-            geoCyl,
-            material,
-            BONES_PER_POSE * this.maxInstances,
-        );
-        // Initialize instanceColor
-        this.bonesMesh.setColorAt(0, COLOR_BASE);
-        this.bonesMesh.count = 0;
-        this.bonesMesh.instanceMatrix.setUsage(THREE.DynamicDrawUsage);
-        this.bonesMesh.instanceColor.setUsage(THREE.DynamicDrawUsage);
-
-        this.container.add(this.bonesMesh);
-    }
-
-    /**
-     * @return {boolean} whether every slot is taken
-     */
-    isFull() {
-        return this.nextInstanceSlot >= this.maxInstances &&
-            this.freeInstanceSlots.length === 0;
-    }
-
-    /**
-     * @return {number} index of the taken slot
-     */
-    takeSlot() {
-        if (this.freeInstanceSlots.length > 0) {
-            return this.freeInstanceSlots.pop();
-        }
-        if (this.nextInstanceSlot >= this.maxInstances) {
-            console.error('out of instances');
-            return 0;
-        }
-        const takenSlot = this.nextInstanceSlot;
-        this.nextInstanceSlot += 1;
-
-        this.jointsMesh.count = JOINTS_PER_POSE * this.nextInstanceSlot;
-        this.bonesMesh.count = BONES_PER_POSE * this.nextInstanceSlot;
-
-        return takenSlot;
-    }
-
-    hideSlot(slot) {
-        let zeros = new THREE.Matrix4().set(
-            0, 0, 0, 0,
-            0, 0, 0, 0,
-            0, 0, 0, 0,
-            0, 0, 0, 1,
-        );
-        for (let i = 0; i < JOINTS_PER_POSE; i++) {
-            this.setJointMatrixAt(slot, i, zeros);
-        }
-        for (let i = 0; i < BONES_PER_POSE; i++) {
-            this.setBoneMatrixAt(slot, i, zeros);
-        }
-    }
-
-    /**
-     * Hides `slot` and adds it to the free list for reuse
-     * @param {number} slot to free up
-     */
-    leaveSlot(slot) {
-        this.freeInstanceSlots.push(slot);
-        this.hideSlot(slot);
-    }
-
-    /**
-     * @param {number} slot - assigned rendering slot
-     * @param {number} index - index of joint within slot
-     * @param {THREE.Vector3} position
-     */
-    setJointMatrixAt(slot, index, matrix) {
-        const offset = slot * JOINTS_PER_POSE + index;
-        this.jointsMesh.setMatrixAt(
-            offset,
-            matrix,
-        );
-        const itemSize = this.jointsMesh.instanceMatrix.itemSize;
-        const updateOffset = offset * itemSize;
-        unionUpdateRange(this.jointsMesh.instanceMatrix, updateOffset, itemSize);
-    }
-
-    /**
-     * @param {number} slot - assigned rendering slot
-     * @param {number} index - index of bone within slot
-     * @param {THREE.Vector3} position
-     */
-    setBoneMatrixAt(slot, index, matrix) {
-        const offset = slot * BONES_PER_POSE + index;
-        this.bonesMesh.setMatrixAt(
-            offset,
-            matrix,
-        );
-
-        const itemSize = this.bonesMesh.instanceMatrix.itemSize;
-        const updateOffset = offset * itemSize;
-        unionUpdateRange(this.bonesMesh.instanceMatrix, updateOffset, itemSize);
-    }
-
-    /**
-     * @param {number} slot - assigned rendering slot
-     * @param {number} index - index of joint within slot
-     * @param {THREE.Color} color
-     */
-    setJointColorAt(slot, index, color) {
-        this.jointsMesh.setColorAt(
-            slot * JOINTS_PER_POSE + index,
-            color,
-        );
-    }
-
-    /**
-     * @param {number} slot - assigned rendering slot
-     * @param {number} index - index of bone within slot
-     * @param {THREE.Color} color
-     */
-    setBoneColorAt(slot, index, color) {
-        this.bonesMesh.setColorAt(
-            slot * BONES_PER_POSE + index,
-            color,
-        );
-    }
-
-    /**
-     * @param {number} slot
-     * @return {Float32Array}
-     */
-    getSlotJointMatrices(slot) {
-        return getSliceSlot(this.jointsMesh.instanceMatrix, slot, JOINTS_PER_POSE);
-    }
-
-    /**
-     * @param {number} slot
-     * @return {Float32Array}
-     */
-    getSlotBoneMatrices(slot) {
-        return getSliceSlot(this.bonesMesh.instanceMatrix, slot, BONES_PER_POSE);
-    }
-
-    /**
-     * @param {number} slot
-     * @param {Float32Array} matrices
-     */
-    setSlotJointMatrices(slot, matrices) {
-        setSliceSlot(this.jointsMesh.instanceMatrix, slot, JOINTS_PER_POSE, matrices);
-    }
-
-    /**
-     * @param {number} slot
-     * @param {Float32Array} matrices
-     */
-    setSlotBoneMatrices(slot, matrices) {
-        setSliceSlot(this.bonesMesh.instanceMatrix, slot, BONES_PER_POSE, matrices);
-    }
-
-
-    /**
-     * @param {number} slot
-     * @return {Float32Array}
-     */
-    getSlotJointColors(slot) {
-        return getSliceSlot(this.jointsMesh.instanceColor, slot, JOINTS_PER_POSE);
-    }
-
-    /**
-     * @param {number} slot
-     * @return {Float32Array}
-     */
-    getSlotBoneColors(slot) {
-        return getSliceSlot(this.bonesMesh.instanceColor, slot, BONES_PER_POSE);
-    }
-
-    /**
-     * @param {number} slot
-     * @param {Float32Array} colors
-     */
-    setSlotJointColors(slot, colors) {
-        setSliceSlot(this.jointsMesh.instanceColor, slot, JOINTS_PER_POSE, colors);
-    }
-
-    /**
-     * @param {number} slot
-     * @param {Float32Array} colors
-     */
-    setSlotBoneColors(slot, colors) {
-        setSliceSlot(this.bonesMesh.instanceColor, slot, BONES_PER_POSE, colors);
-    }
-
-    addToScene(container) {
-        if (container) {
-            container.add(this.container);
-        } else {
-            realityEditor.gui.threejsScene.addToScene(this.container);
-        }
-    }
-
-    /**
-     * Removes from container and disposes resources
-     */
-    removeFromScene(container) {
-        if (container) {
-            container.remove(this.container);
-        } else {
-            realityEditor.gui.threejsScene.removeFromScene(this.container);
-        }
-        this.bonesMesh.dispose();
-        this.jointsMesh.dispose();
-    }
-
-    markNeedsUpdate() {
-        this.markMatrixNeedsUpdate();
-        this.markColorNeedsUpdate();
-    }
-
-    markMatrixNeedsUpdate() {
-        this.jointsMesh.instanceMatrix.needsUpdate = true;
-        this.bonesMesh.instanceMatrix.needsUpdate = true;
-    }
-
-    markColorNeedsUpdate() {
-        this.jointsMesh.instanceColor.needsUpdate = true;
-        this.bonesMesh.instanceColor.needsUpdate = true;
-    }
-}
-
-/**
- * A single 3d skeleton rendered in a HumanPoseRenderer's slot
- */
-class HumanPoseRenderInstance {
-    /**
-     * @param {HumanPoseRenderer} renderer
-     * @param {string} id - Unique identifier of human pose being rendered
-     */
-    constructor(renderer, id) {
-        this.renderer = renderer;
-        this.id = id;
-        this.updated = true;
-        this.visible = true;
-        this.jointPositions = [];
-        for (let i = 0; i < JOINTS_PER_POSE; i++) {
-            this.jointPositions.push(new THREE.Vector3(0, 0, 0));
-        }
-        this.previousMatrices = null;
-        this.overallRebaScore = 1;
-        this.colorOptions = [];
-        this.slot = -1;
-        this.add();
-    }
-
-    /**
-     * Occupies a slot on the renderer, uploading initial values
-     * @param {number?} slot - manually assigned slot, taken from renderer
-     * otherwise
-     */
-    add(slot) {
-        if (typeof slot === 'number') {
-            this.slot = slot;
-        } else {
-            this.slot = this.renderer.takeSlot();
-        }
-
-        for (const [i, _jointId] of Object.values(JOINTS).entries()) {
-            this.renderer.setJointColorAt(this.slot, i, COLOR_BASE);
-        }
-
-        for (const [i, _boneName] of Object.keys(JOINT_CONNECTIONS).entries()) {
-            this.renderer.setBoneColorAt(this.slot, i, COLOR_BASE);
-        }
-    }
-
-    /**
-     * @param {string} jointId - from utils.JOINTS
-     * @param {THREE.Vector3} position
-     */
-    setJointPosition(jointId, position) {
-        const index = JOINT_TO_INDEX[jointId];
-        this.jointPositions[index] = position;
-
-        this.renderer.setJointMatrixAt(
-            this.slot,
-            index,
-            new THREE.Matrix4().makeTranslation(
-                position.x,
-                position.y,
-                position.z,
-            ),
-        );
-    }
-
-    /**
-     * @return {THREE.Vector3}
-     */
-    getJointPosition(jointId) {
-        const index = JOINT_TO_INDEX[jointId];
-        return this.jointPositions[index];
-    }
-
-    /**
-     * @param {Array<String>} jointIds
-     * @return {{x: number, y: number, z: number}} Average position of all
-     *         joints listed in jointIds
-     */
-    averageJointPositions(jointIds) {
-        let avg = {x: 0, y: 0, z: 0};
-        for (let jointId of jointIds) {
-            let jointPos = this.getJointPosition(jointId);
-            avg.x += jointPos.x;
-            avg.y += jointPos.y;
-            avg.z += jointPos.z;
-        }
-        avg.x /= jointIds.length;
-        avg.y /= jointIds.length;
-        avg.z /= jointIds.length;
-        return avg;
-    }
-
-    /**
-     * Sets all matrices in our assigned renderer slot based on the current
-     * state of this instance
-     */
-    setAllMatrices() {
-        for (let i = 0; i < this.jointPositions.length; i++) {
-            let position = this.jointPositions[i];
-            this.renderer.setJointMatrixAt(
-                this.slot,
-                i,
-                new THREE.Matrix4().makeTranslation(
-                    position.x,
-                    position.y,
-                    position.z,
-                ),
-            );
-        }
-
-        this.updateBonePositions();
-    }
-
-    /**
-     * Updates bone (stick between joints) positions based on joint
-     * positions.
-     */
-    updateBonePositions() {
-        for (let boneName of Object.keys(JOINT_CONNECTIONS)) {
-            const boneIndex = BONE_TO_INDEX[boneName];
-            let jointA = this.getJointPosition(JOINT_CONNECTIONS[boneName][0]);
-            let jointB = this.getJointPosition(JOINT_CONNECTIONS[boneName][1]);
-
-            let pos = new THREE.Vector3(
-                (jointA.x + jointB.x) / 2,
-                (jointA.y + jointB.y) / 2,
-                (jointA.z + jointB.z) / 2,
-            );
-
-            let diff = new THREE.Vector3(jointB.x - jointA.x, jointB.y - jointA.y,
-                jointB.z - jointA.z);
-            let scale = new THREE.Vector3(1, diff.length() / SCALE, 1);
-            diff.normalize();
-
-            let rot = new THREE.Quaternion();
-            rot.setFromUnitVectors(new THREE.Vector3(0, 1, 0),
-                                   diff);
-
-            // bone.lookAt(this.container.localToWorld(localTarget));
-            // bone.rotateX(Math.PI / 2);
-            let mat = new THREE.Matrix4();
-            mat.compose(pos, rot, scale);
-
-            this.renderer.setBoneMatrixAt(this.slot, boneIndex, mat);
-        }
-
-        if (!RENDER_CONFIDENCE_COLOR) {
-            annotateHumanPoseRenderer(this);
-        }
-    }
-
-    setOverallRebaScore(score) {
-        this.overallRebaScore = score;
-    }
-
-    /**
-     * Annotates bone using material based on boneColor
-     * @param {string} boneName
-     * @param {number} boneColor
-     */
-    setBoneRebaColor(boneName, boneColor) {
-        if (typeof BONE_TO_INDEX[boneName] === 'undefined') {
-            return;
-        }
-
-        const boneIndex = BONE_TO_INDEX[boneName];
-        const joint = JOINT_CONNECTIONS[boneName][1];
-        const jointIndex = JOINT_TO_INDEX[joint];
-
-        let color = COLOR_BASE;
-        if (boneColor === 0) {
-            color = COLOR_GREEN;
-        } else if (boneColor === 1) {
-            color = COLOR_YELLOW;
-        } else if (boneColor === 2) {
-            color = COLOR_RED;
-        }
-        this.renderer.setBoneColorAt(this.slot, boneIndex, color);
-        this.renderer.setJointColorAt(this.slot, jointIndex, color);
-    }
-
-    /**
-     * @param {number} confidence in range [0,1]
-     */
-    setJointConfidenceColor(jointId, confidence) {
-        if (typeof JOINT_TO_INDEX[jointId] === 'undefined') {
-            return;
-        }
-        const jointIndex = JOINT_TO_INDEX[jointId];
-
-        let baseColorHSL = {};
-        this.baseColor.getHSL(baseColorHSL);
-
-        baseColorHSL.l = baseColorHSL.l * confidence;
-
-        let color = new THREE.Color();
-        color.setHSL(baseColorHSL.h, baseColorHSL.s, baseColorHSL.l);
-
-        this.renderer.setJointColorAt(this.slot, jointIndex, color);
-    }
-
-    setColorOption(colorOptionIndex) {
-        this.renderer.setSlotBoneColors(this.slot, this.colorOptions[colorOptionIndex].boneColors);
-        this.renderer.setSlotJointColors(this.slot, this.colorOptions[colorOptionIndex].jointColors);
-    }
-
-    setVisible(visible) {
-        if (this.visible === visible) {
-            return;
-        }
-
-        if (this.visible) {
-            if (!this.previousMatrices) {
-                this.previousMatrices = {
-                    boneMatrices: this.renderer.getSlotBoneMatrices(this.slot),
-                    jointMatrices: this.renderer.getSlotJointMatrices(this.slot),
-                };
-            }
-            this.renderer.hideSlot(this.slot);
-        } else {
-            this.renderer.setSlotBoneMatrices(this.slot, this.previousMatrices.boneMatrices);
-            this.renderer.setSlotJointMatrices(this.slot, this.previousMatrices.jointMatrices);
-        }
-        this.visible = visible;
-    }
-
-    getOverallRebaScoreHue() {
-        let hueReba = 140 - (this.overallRebaScore - 1) * 240 / 11;
-        if (isNaN(hueReba)) {
-            hueReba = 120;
-        }
-        hueReba = (Math.min(Math.max(hueReba, -30), 120) + 360) % 360;
-        return hueReba;
-    }
-
-    cloneToRenderer(newRenderer, startingColorOption) {
-        let clone = new HumanPoseRenderInstance(newRenderer, this.id);
-
-        for (const jointId of Object.keys(JOINT_TO_INDEX)) {
-            clone.setJointPosition(jointId, this.getJointPosition(jointId));
-        }
-        // TODO would be significantly faster to use the bulk set methods
-        clone.updateBonePositions();
-
-        let colorRainbow = new THREE.Color();
-        colorRainbow.setHSL(((Date.now() / 5) % 360) / 360, 1, 0.5);
-
-        let hueReba = this.getOverallRebaScoreHue();
-        // let alphaReba = 0.3 + 0.3 * (poseRenderer.overallRebaScore - 1) / 11;
-        let colorReba = new THREE.Color();
-        colorReba.setHSL(hueReba / 360, 1, 0.5);
-
-        let boneColors = this.renderer.getSlotBoneColors(this.slot);
-        let jointColors = this.renderer.getSlotJointColors(this.slot);
-
-        let boneColorsRainbow = boneColors.slice(0);
-        let jointColorsRainbow = jointColors.slice(0);
-
-        let boneColorsReba = boneColors.slice(0);
-        let jointColorsReba = jointColors.slice(0);
-
-        for (let i = 0; i < boneColors.length / 3; i++) {
-            colorRainbow.toArray(boneColorsRainbow, i * 3);
-            colorReba.toArray(boneColorsReba, i * 3);
-        }
-        for (let i = 0; i < jointColors.length / 3; i++) {
-            colorRainbow.toArray(jointColorsRainbow, i * 3);
-            colorReba.toArray(jointColorsReba, i * 3);
-        }
-
-        clone.colorOptions.push({
-            boneColors: boneColorsReba,
-            jointColors: jointColorsReba,
-        });
-
-        clone.colorOptions.push({
-            boneColors,
-            jointColors,
-        });
-
-        clone.colorOptions.push({
-            boneColors: boneColorsRainbow,
-            jointColors: jointColorsRainbow,
-        });
-
-        clone.setColorOption(startingColorOption);
-
-        return clone;
-    }
-
-    /**
-     * Removes from container and disposes resources
-     */
-    remove() {
-        this.renderer.leaveSlot(this.slot);
-    }
-}
 
 const AnimationMode = {
     ONE: 'one',
